requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pypi_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.3.1=pypi_0
easydict=1.11=pypi_0
ftfy=6.1.1=pypi_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.59.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.7.0=pypi_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=13.2.0=h807b86a_2
libgomp=13.2.0=h807b86a_2
libstdcxx-ng=13.2.0=h7e041cc_2
libzlib=1.2.13=hd590300_5
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
ncurses=6.4=hcb278e6_0
numpy=1.21.6=pypi_0
oauthlib=3.2.2=pypi_0
openssl=1.0.2u=h516909a_0
pillow=9.5.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
prettytable=3.7.0=pypi_0
protobuf=3.20.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
python=3.7.0=hd21baee_1006
pyyaml=6.0.1=pypi_0
readline=7.0=hf8c457e_1001
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
setuptools=58.0.4=pypi_0
six=1.16.0=pypi_0
sqlite=3.28.0=h8b20d00_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
timm=0.6.7=pypi_0
tk=8.6.13=h2797004_0
torch=1.9.0+cu111=pypi_0
torchvision=0.10.0+cu111=pypi_0
tqdm=4.66.1=pypi_0
typing-extensions=4.7.1=pypi_0
urllib3=2.0.7=pypi_0
wcwidth=0.2.8=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.41.2=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
zipp=3.15.0=pypi_0
zlib=1.2.13=hd590300_5
